// src/generated/types.gen.ts
export type ClubActivity = {
  athlete?: MetaAthlete
  /**
   * The name of the activity
   */
  name?: string
  /**
   * The activity's distance, in meters
   */
  distance?: number
  /**
   * The activity's moving time, in seconds
   */
  moving_time?: number
  /**
   * The activity's elapsed time, in seconds
   */
  elapsed_time?: number
  /**
   * The activity's total elevation gain.
   */
  total_elevation_gain?: number
  type?: ActivityType
  sport_type?: SportType
  /**
   * The activity's workout type
   */
  workout_type?: number
}

export type DetailedActivity = SummaryActivity & {
  /**
   * The description of the activity
   */
  description?: string
  photos?: PhotosSummary
  gear?: SummaryGear
  /**
   * The number of kilocalories consumed during this activity
   */
  calories?: number
  segment_efforts?: Array<DetailedSegmentEffort>
  /**
   * The name of the device used to record the activity
   */
  device_name?: string
  /**
   * The token used to embed a Strava activity
   */
  embed_token?: string
  /**
   * The splits of this activity in metric units (for runs)
   */
  splits_metric?: Array<Split>
  /**
   * The splits of this activity in imperial units (for runs)
   */
  splits_standard?: Array<Split>
  laps?: Array<Lap>
  best_efforts?: Array<DetailedSegmentEffort>
}
// ...
export type ActivityStats = {
  /**
   * The longest distance ridden by the athlete.
   */
  biggest_ride_distance?: number
  /**
   * The highest climb ridden by the athlete.
   */
  biggest_climb_elevation_gain?: number
  recent_ride_totals?: ActivityTotal
  recent_run_totals?: ActivityTotal
  recent_swim_totals?: ActivityTotal
  ytd_ride_totals?: ActivityTotal
  ytd_run_totals?: ActivityTotal
  ytd_swim_totals?: ActivityTotal
  all_ride_totals?: ActivityTotal
  all_run_totals?: ActivityTotal
  all_swim_totals?: ActivityTotal
}
// ...
export type ClubAthlete = {
  /**
   * Resource state, indicates level of detail. Possible values: 1 -> "meta", 2 -> "summary", 3 -> "detail"
   */
  resource_state?: number
  /**
   * The athlete's first name.
   */
  firstname?: string
  /**
   * The athlete's last initial.
   */
  lastname?: string
  /**
   * The athlete's member status.
   */
  member?: string
  /**
   * Whether the athlete is a club admin.
   */
  admin?: boolean
  /**
   * Whether the athlete is club owner.
   */
  owner?: boolean
}

export type DetailedAthlete = SummaryAthlete & {
  /**
   * The athlete's follower count.
   */
  follower_count?: number
  /**
   * The athlete's friend count.
   */
  friend_count?: number
  /**
   * The athlete's preferred unit system.
   */
  measurement_preference?: 'feet' | 'meters'
  /**
   * The athlete's FTP (Functional Threshold Power).
   */
  ftp?: number
  /**
   * The athlete's weight.
   */
  weight?: number
  /**
   * The athlete's clubs.
   */
  clubs?: Array<SummaryClub>
  /**
   * The athlete's bikes.
   */
  bikes?: Array<SummaryGear>
  /**
   * The athlete's shoes.
   */
  shoes?: Array<SummaryGear>
}

export type MetaAthlete = {
  /**
   * The unique identifier of the athlete
   */
  id?: number
}
// ...
export type DetailedClub = SummaryClub & {
  /**
   * The membership status of the logged-in athlete.
   */
  membership?: 'member' | 'pending'
  /**
   * Whether the currently logged-in athlete is an administrator of this club.
   */
  admin?: boolean
  /**
   * Whether the currently logged-in athlete is the owner of this club.
   */
  owner?: boolean
  /**
   * The number of athletes in the club that the logged-in athlete follows.
   */
  following_count?: number
}
// ...
export type Comment = {
  /**
   * The unique identifier of this comment
   */
  id?: number
  /**
   * The identifier of the activity this comment is related to
   */
  activity_id?: number
  /**
   * The content of the comment
   */
  text?: string
  athlete?: SummaryAthlete
  /**
   * The time at which this comment was created.
   */
  created_at?: string
}
// ...
export type DetailedGear = SummaryGear & {
  /**
   * The gear's brand name.
   */
  brand_name?: string
  /**
   * The gear's model name.
   */
  model_name?: string
  /**
   * The gear's frame type (bike only).
   */
  frame_type?: number
  /**
   * The gear's description.
   */
  description?: string
}
// ...
export type DetailedSegment = SummarySegment & {
  /**
   * The time at which the segment was created.
   */
  created_at?: string
  /**
   * The time at which the segment was last updated.
   */
  updated_at?: string
  /**
   * The segment's total elevation gain.
   */
  total_elevation_gain?: number
  map?: PolylineMap
  /**
   * The total number of efforts for this segment
   */
  effort_count?: number
  /**
   * The number of unique athletes who have an effort for this segment
   */
  athlete_count?: number
  /**
   * Whether this segment is considered hazardous
   */
  hazardous?: boolean
  /**
   * The number of stars for this segment
   */
  star_count?: number
}

export type ExplorerResponse = {
  /**
   * The set of segments matching an explorer request
   */
  segments?: Array<ExplorerSegment>
}
// ...
export type DetailedSegmentEffort = SummarySegmentEffort & {
  /**
   * The name of the segment on which this effort was performed
   */
  name?: string
  activity?: MetaActivity
  athlete?: MetaAthlete
  /**
   * The effort's moving time
   */
  moving_time?: number
  /**
   * The start index of this effort in its activity's stream
   */
  start_index?: number
  /**
   * The end index of this effort in its activity's stream
   */
  end_index?: number
  /**
   * The effort's average cadence
   */
  average_cadence?: number
  /**
   * The average wattage of this effort
   */
  average_watts?: number
  /**
   * For riding efforts, whether the wattage was reported by a dedicated recording device
   */
  device_watts?: boolean
  /**
   * The heart heart rate of the athlete during this effort
   */
  average_heartrate?: number
  /**
   * The maximum heart rate of the athlete during this effort
   */
  max_heartrate?: number
  segment?: SummarySegment
  /**
   * The rank of the effort on the global leaderboard if it belongs in the top 10 at the time of upload
   */
  kom_rank?: number
  /**
   * The rank of the effort on the athlete's leaderboard if it belongs in the top 3 at the time of upload
   */
  pr_rank?: number
  /**
   * Whether this effort should be hidden when viewed within an activity
   */
  hidden?: boolean
}
// ...
export type ActivityZone = {
  score?: number
  distribution_buckets?: TimedZoneDistribution
  type?: 'heartrate' | 'power'
  sensor_based?: boolean
  points?: number
  custom_zones?: boolean
  max?: number
}

export type HeartRateZoneRanges = {
  /**
   * Whether the athlete has set their own custom heart rate zones
   */
  custom_zones?: boolean
  zones?: ZoneRanges
}

export type PowerZoneRanges = {
  zones?: ZoneRanges
}

/**
 * Stores the exclusive ranges representing zones and the time spent in each.
 */
export type TimedZoneDistribution = Array<TimedZoneRange>

/**
 * A union type representing the time spent in a given zone.
 */
export type TimedZoneRange = ZoneRange & {
  /**
   * The number of seconds spent in this zone
   */
  time?: number
}

export type ZoneRange = {
  /**
   * The minimum value in the range.
   */
  min?: number
  /**
   * The maximum value in the range.
   */
  max?: number
}

export type ZoneRanges = Array<ZoneRange>

export type Zones = {
  heart_rate?: HeartRateZoneRanges
  power?: PowerZoneRanges
}
// ...
export type GetEffortsBySegmentIdData = {
  body?: never
  path?: never
  query: {
    /**
     * The identifier of the segment.
     */
    segment_id: number
    /**
     * ISO 8601 formatted date time.
     */
    start_date_local?: string
    /**
     * ISO 8601 formatted date time.
     */
    end_date_local?: string
    /**
     * Number of items per page. Defaults to 30.
     */
    per_page?: number
  }
  url: '/segment_efforts'
}
// ...
export type ExploreSegmentsData = {
  body?: never
  path?: never
  query: {
    /**
     * The latitude and longitude for two points describing a rectangular boundary for the search: [southwest corner latitutde, southwest corner longitude, northeast corner latitude, northeast corner longitude]
     */
    bounds: [number, number, number, number]
    /**
     * Desired activity type.
     */
    activity_type?: 'running' | 'riding'
    /**
     * The minimum climbing category.
     */
    min_cat?: number
    /**
     * The maximum climbing category.
     */
    max_cat?: number
  }
  url: '/segments/explore'
}
// ...
export type CreateActivityData = {
  body: {
    /**
     * The name of the activity.
     */
    name: string
    /**
     * Type of activity. For example - Run, Ride etc.
     */
    type?: string
    /**
     * Sport type of activity. For example - Run, MountainBikeRide, Ride, etc.
     */
    sport_type: string
    /**
     * ISO 8601 formatted date time.
     */
    start_date_local: string
    /**
     * In seconds.
     */
    elapsed_time: number
    /**
     * Description of the activity.
     */
    description?: string
    /**
     * In meters.
     */
    distance?: number
    /**
     * Set to 1 to mark as a trainer activity.
     */
    trainer?: number
    /**
     * Set to 1 to mark as commute.
     */
    commute?: number
  }
  path?: never
  query?: never
  url: '/activities'
}
// ...
export type GetLoggedInAthleteActivitiesData = {
  body?: never
  path?: never
  query?: {
    /**
     * An epoch timestamp to use for filtering activities that have taken place before a certain time.
     */
    before?: number
    /**
     * An epoch timestamp to use for filtering activities that have taken place after a certain time.
     */
    after?: number
    /**
     * Page number. Defaults to 1.
     */
    page?: number
    /**
     * Number of items per page. Defaults to 30.
     */
    per_page?: number
  }
  url: '/athlete/activities'
}
// ...
export type GetCommentsByActivityIdData = {
  body?: never
  path: {
    /**
     * The identifier of the activity.
     */
    id: number
  }
  query?: {
    /**
     * Deprecated. Prefer to use after_cursor.
     */
    page?: number
    /**
     * Deprecated. Prefer to use page_size.
     */
    per_page?: number
    /**
     * Number of items per page. Defaults to 30.
     */
    page_size?: number
    /**
     * Cursor of the last item in the previous page of results, used to request the subsequent page of results.  When omitted, the first page of results is fetched.
     */
    after_cursor?: string
  }
  url: '/activities/{id}/comments'
}
// ...
export type GetKudoersByActivityIdData = {
  body?: never
  path: {
    /**
     * The identifier of the activity.
     */
    id: number
  }
  query?: {
    /**
     * Page number. Defaults to 1.
     */
    page?: number
    /**
     * Number of items per page. Defaults to 30.
     */
    per_page?: number
  }
  url: '/activities/{id}/kudos'
}
// ...
export type GetClubMembersByIdData = {
  body?: never
  path: {
    /**
     * The identifier of the club.
     */
    id: number
  }
  query?: {
    /**
     * Page number. Defaults to 1.
     */
    page?: number
    /**
     * Number of items per page. Defaults to 30.
     */
    per_page?: number
  }
  url: '/clubs/{id}/members'
}
// ...
export type GetClubAdminsByIdData = {
  body?: never
  path: {
    /**
     * The identifier of the club.
     */
    id: number
  }
  query?: {
    /**
     * Page number. Defaults to 1.
     */
    page?: number
    /**
     * Number of items per page. Defaults to 30.
     */
    per_page?: number
  }
  url: '/clubs/{id}/admins'
}
// ...
export type GetClubActivitiesByIdData = {
  body?: never
  path: {
    /**
     * The identifier of the club.
     */
    id: number
  }
  query?: {
    /**
     * Page number. Defaults to 1.
     */
    page?: number
    /**
     * Number of items per page. Defaults to 30.
     */
    per_page?: number
  }
  url: '/clubs/{id}/activities'
}
// ...
export type CreateUploadData = {
  body?: {
    /**
     * The uploaded file.
     */
    file?: Blob | File
    /**
     * The desired name of the resulting activity.
     */
    name?: string
    /**
     * The desired description of the resulting activity.
     */
    description?: string
    /**
     * Whether the resulting activity should be marked as having been performed on a trainer.
     */
    trainer?: string
    /**
     * Whether the resulting activity should be tagged as a commute.
     */
    commute?: string
    /**
     * The format of the uploaded file.
     */
    data_type?: 'fit' | 'fit.gz' | 'tcx' | 'tcx.gz' | 'gpx' | 'gpx.gz'
    /**
     * The desired external identifier of the resulting activity.
     */
    external_id?: string
  }
  path?: never
  query?: never
  url: '/uploads'
}